requirements.txt
python-dotenv
llama-index
openai
judgeval